requirements.txt
streamlit==1.36.0
pandas>=2.0.0
yfinance>=0.2.27
matplotlib>=3.7.0
statsmodels>=0.14.0
pyyaml>=6.0
requests>=2.31.0
openpyxl>=3.1.2
